requirements.txt
numpy == 1.21.2
scipy == 1.7.3
pandas == 1.3.4
torch == 1.10.0
scikit-learn == 1.0.1
wfdb == 4.1.0
mlflow == 1.27.0
matplotlib == 3.4.3
jupyterlab == 3.5.3
PyYAML == 6.0